requirements.txt
django==3.2
pymysql==1.1.0
cryptography==41.0.1
Pillow==9.5.0
django-simple-captcha==0.5.18